@@ -8,72 +8,70 @@
 
 cmake_minimum_required (VERSION 2.6)
 ENABLE_LANGUAGE(CXX)
-#set(CMAKE_BUILD_TYPE Release)
+set(CMAKE_BUILD_TYPE Release)
 
 project (DeathKnell)
 set(PROJECT_SRC ${DeathKnell_SOURCE_DIR}/src)
 set(DIR_UNIT_TEST ${DeathKnell_SOURCE_DIR}/test)
-set(DIR_3RDPARTY ${DeathKnell_SOURCE_DIR}/3rdparty)
+set(DIR_3RDPARTY ${DeathKnell_SOURCE_DIR}/thirdparty)
 
 
 MESSAGE(" PROJECT_SRC = : ${PROJECT_SRC}")
 include_directories(${PROJECT_SRC})
 SET(ACTIVE_CPP0xx_DIR "Release")
 
-find_library(G2LOG lib_g2log PATHS /usr/local/probe/lib )
-find_library(BOOST_THREAD boost_thread PATHS /usr/local/probe/lib )
-find_library(BOOST_SYSTEM boost_thread PATHS /usr/local/probe/lib )
-find_library(TCMALLOC tcmalloc PATHS /usr/local/probe/lib )
+find_library(G3LOG g3logger PATHS /usr/local/probe/lib )
+set(LIBS ${G3LOG})
 
-#cmake -DCMAKE_CXX_COMPILER=clang++ ..
-  # WARNING: If Clang for Linux does not work with full C++11 support it might be your
-  # installation that is faulty. When I tested Clang on Ubuntu I followed the following
-  # description
-  # 1) http://kjellkod.wordpress.com/2013/09/23/experimental-g2log-with-clang/
-  # 2) https://github.com/maidsafe/MaidSafe/wiki/Hacking-with-Clang-llvm-abi-and-llvm-libc
-IF ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "Clang")
-   MESSAGE("")
-   MESSAGE("cmake for Clang ")
-   if (APPLE)
-       set(PLATFORM_LINK_LIBRIES c++abi)
-   else()
-       set(PLATFORM_LINK_LIBRIES rt c++abi)
-   endif()
-  set(CMAKE_CXX_FLAGS "-Wall -std=c++11 -pthread -stdlib=libc++ -Wunused -D_GLIBCXX_USE_NANOSLEEP")
+find_library(G3FILE_ROTATE g3logrotate PATHS /usr/local/probe/lib)
+list(APPEND LIBS ${G3FILE_ROTATE})
 
-ELSEIF("${CMAKE_CXX_COMPILER_ID}" STREQUAL "GNU")
+find_library(FILEIO FileIO PATHS /usr/local/probe/lib)
+list(APPEND LIBS ${FILEIO})
+
+
+IF ("${CMAKE_CXX_COMPILER_ID}" STREQUAL "GNU")
    MESSAGE("")
    MESSAGE("cmake for GCC ")
-   if (APPLE)
-       set(CMAKE_CXX_FLAGS "-Wall -Wunused -std=c++11 -pthread -D_GLIBCXX_USE_NANOSLEEP")
-   else()
-       set(PLATFORM_LINK_LIBRIES rt)
-       set(CMAKE_CXX_FLAGS "-Wall -rdynamic -Wunused -std=c++11 -pthread -D_GLIBCXX_USE_NANOSLEEP")
-   endif()
-
-ELSEIF(MSVC)
-      # VC11 bug: http://code.google.com/p/googletest/issues/detail?id=408
-      # add_definition(-D_VARIADIC_MAX=10)
-      # https://github.com/anhstudios/swganh/pull/186/files
-      ADD_DEFINITIONS (/D_VARIADIC_MAX=10)
-      MESSAGE(STATUS "- MSVC: Set variadic max to 10 for MSVC compatibility")
-      # Remember to set set target properties if using GTEST similar to done below on target "unit_test"
-      # "set_target_properties(unit_test PROPERTIES COMPILE_DEFINITIONS "GTEST_USE_OWN_TR1_TUPLE=0")
-   MESSAGE("")
-   MESSAGE("Windows: Please run the command [cmake -DCMAKE_BUILD_TYPE=Release -G \"Visual Studio 11\" ..]")
-   MESSAGE("if cmake finishes OK, do 'msbuild TestOfDeathKnell.sln /p:Configuration=Release'")
-   MESSAGE("then run 'Release\\TestOfDeathKnell.exe'")
-   MESSAGE("")
+   SET(PLATFORM_LINK_LIBRIES rt)
+   SET(CMAKE_CXX_FLAGS "-Wall -rdynamic -Wunused -std=c++14 -pthread -D_GLIBCXX_USE_NANOSLEEP")
 ENDIF()
 
 
 # GENERIC STEPS
-file(GLOB SRC_FILES ${PROJECT_SRC}/*.h ${PROJECT_SRC}/*.hpp ${PROJECT_SRC}/*.cpp ${PROJECT_SRC}/*.ipp)
+MESSAGE("PROJECT SRC: ${PROJECT_SRC}")
+FILE(GLOB SRC_FILES ${PROJECT_SRC}/*.h ${PROJECT_SRC}/*.hpp ${PROJECT_SRC}/*.cpp) 
  
+
+# Setup Library name
+# Name for the library
+SET(LIBRARY_TO_BUILD DeathKnell)
+
+SET(MAJOR_VERSION 1)
+
+# Find the current version if not already set
+# =============================================
+# workaround due to cmake 2.8 or otherwise where variables cannot 
+# easily be detected for existance
+# i.e the following does NOT work 
+# IF (NOT DEFINED ${VERSION})
+#
+IF (${VERSION}.x STREQUAL ".x")   
+   MESSAGE("Extracting git software version")
+   execute_process(COMMAND bash "-c" "git rev-list --branches HEAD | wc -l | tr -d ' ' | tr -d '\n'" OUTPUT_VARIABLE GIT_VERSION)
+   SET(MINOR_VERSION ${GIT_VERSION})
+   SET(BUILD_NUMBER 0)
+   SET(VERSION ${MAJOR_VERSION}.${MINOR_VERSION}.${BUILD_NUMBER}) 
+ENDIF()
+MESSAGE("Software Version: ${VERSION}")
+
+
 # Create the DeathKnell library
 include_directories(${PROJECT_SRC})
-add_library(DeathKnell ${SRC_FILES})
-set_target_properties(DeathKnell PROPERTIES LINKER_LANGUAGE CXX)
+add_library(${LIBRARY_TO_BUILD} SHARED  ${SRC_FILES})
+SET(DeathKnell_VERSION_STRING ${VERSION})
+SET_TARGET_PROPERTIES(${LIBRARY_TO_BUILD} PROPERTIES LINKER_LANGUAGE CXX SOVERSION ${VERSION})
+TARGET_LINK_LIBRARIES(${LIBRARY_TO_BUILD} ${LIBS})
 
 
 
@@ -87,19 +85,22 @@ add_library(gtest_170_lib ${GTEST_DIR}/src/gtest-all.cc)
 set_target_properties(gtest_170_lib PROPERTIES COMPILE_DEFINITIONS "GTEST_HAS_RTTI=0")
 enable_testing(true)
 
-
-
-# obs see this: http://stackoverflow.com/questions/9589192/how-do-i-change-the-number-of-template-arguments-supported-by-msvcs-stdtupl
-# and this: http://stackoverflow.com/questions/2257464/google-test-and-visual-studio-2010-rc
-#include_directories(${DIR_UNIT_TEST})
 include_directories(test)
-#file(GLOB TEST_SRC_FILES ${DIR_UNIT_TEST}"/*.h" ${DIR_UNIT_TEST}"/*.hpp" ${DIR_UNIT_TEST}"/*.cpp" ${DIR_UNIT_TEST}"/*.ipp")
-  file(GLOB TEST_SRC_FILES "test/*.cpp")
+include_directories(${PROJECT_SRC})
+file(GLOB TEST_SRC_FILES "test/*.cpp")
 
-  add_executable(UnitTestRunner 3rdparty/test_main.cpp ${TEST_SRC_FILES} )
-  target_link_libraries(UnitTestRunner gtest_170_lib boost_thread boost_system lib_g2log stdc++ tcmalloc FileIO ${PLATFORM_LINK_LIBRIES} DeathKnell -Wl,-rpath,. -Wl,-rpath,/usr/local/probe/lib  -Wl,-rpath,/usr/local/probe/lib64 )
-  set_target_properties(${test} PROPERTIES COMPILE_DEFINITIONS "GTEST_HAS_TR1_TUPLE=0")
-  set_target_properties(${test} PROPERTIES COMPILE_DEFINITIONS "GTEST_HAS_RTTI=0")
-IF( NOT(MSVC))
-  set_target_properties(${test} PROPERTIES COMPILE_FLAGS "-isystem -pthread ")
-ENDIF( NOT(MSVC))
+SET(CMAKE_CXX_FLAGS "-Wall -Wno-sign-compare -rdynamic -Wunused -std=c++14 -pthread -D_GLIBCXX_USE_NANOSLEEP")
+add_executable(UnitTestRunner thirdparty/test_main.cpp ${TEST_SRC_FILES} )
+target_link_libraries(UnitTestRunner ${LIBRARY_TO_BUILD} gtest_170_lib ${LIBS})
+set_target_properties(${test} PROPERTIES COMPILE_FLAGS "-isystem -pthread ")
+
+
+IF(${CMAKE_SYSTEM_NAME} MATCHES "Linux" OR ${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
+   FILE(GLOB HEADER_FILES ${PROJECT_SRC}/*.h)
+   # ==========================================================================
+   #   Package handling is done AFTER all other CMake setup
+   #   usage:   make package
+   #   Check the output result and install accordingly.
+   # ==========================================================================
+   INCLUDE (${PROJECT_SRC}/../CPackLists.txt)
+ENDIF()
